Converted http://db.cs.berkeley.edu/ to pure HTML.
Recently, there's been a desire to refresh the Berkeley database group
website [1], as it hasn't been updated in quite some time. Moreover,
we'd like to make the website as easy to update as possible, so that the
website doesn't go stagnant again.

Previously, the website was written using Jekyll. This can deter folks
from updating the website because in order to do so, you first have to
learn Jekyll. For such a simple website, Jekyll is largely unnecessary,
so we're gutting Jekyll and implementing the website in pure HTML. If we
ever switch over to a fancier website, then maybe we can switch back to
something like Jekyll.

[1]: http://db.cs.berkeley.edu/